moved hypernym methods to LanguageTool
