+++ WSD/src/cass/languageTool/LanguageTool.java
@@ -1,6 +1,7 @@
 package cass.languageTool;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import java.util.Set;
 
@@ -11,12 +12,16 @@ import cass.languageTool.tokenizer.*;
 import cass.languageTool.wordNet.*;
 
 import java.util.ListIterator;
+import java.util.Random;
 
 public class LanguageTool implements I_Lemma, I_Tokenizer, I_WordNet, I_PartOfSpeech {
 	private I_WordNet wordNet;
 	private I_Tokenizer tokenizer;
 	private I_Lemma lemmatizer;
 	private I_PartOfSpeech pos;
+	
+	private Random rand = new Random();
+
 	
 	public LanguageTool(Language language) {
 		switch (language) {
@@ -82,6 +87,40 @@ public class LanguageTool implements I_Lemma, I_Tokenizer, I_WordNet, I_PartOfSp
 		return pos.getPOStag(word);
 	}
 
-	
+	public int getHypernymDistanceScore(CASSWordSense sense1, CASSWordSense sense2) {
 		
+		List<CASSWordSense> ancestors1 = getHypernymAncestors(sense1);
+		List<CASSWordSense> ancestors2 = getHypernymAncestors(sense2);
+		
+		Collections.reverse(ancestors1);
+		Collections.reverse(ancestors2);
+		
+		// common ancestor has same ancestor path in both lists
+		int depthOfCommonAncestor = 0;
+		while(ancestors1.get(depthOfCommonAncestor+1).getId() != ancestors2.get(depthOfCommonAncestor+1).getId()) {
+			depthOfCommonAncestor++;
+		}
+		
+		int distanceFromAncestorToS1 = ancestors1.size() - depthOfCommonAncestor;
+		int distanceFromAncestorToS2 = ancestors2.size() - depthOfCommonAncestor;
+		
+		int distanceBetweenSenses = distanceFromAncestorToS1 + distanceFromAncestorToS2;
+		
+		return distanceBetweenSenses;
+	}
+	
+	private List<CASSWordSense> getHypernymAncestors(CASSWordSense sense) {
+		List<CASSWordSense> ancestors = new ArrayList<CASSWordSense>();
+		int size, randomIndex;
+		while (sense.getId() != "entity") {
+			Set<CASSWordSense> hypernyms = getHypernyms(sense);
+			size = hypernyms.size();
+			randomIndex = rand.nextInt(size);
+			CASSWordSense[] hypernymArray = new CASSWordSense[size];
+			hypernyms.toArray(hypernymArray);
+			sense = hypernymArray[randomIndex];
+			ancestors.add(sense);
+		}
+		return ancestors;
+	}	
 }
